:art: Split repo handling to a separate PackageManager class.
